refactor: 네이밍 및 컨트롤러(+URL) 수정 및 빈공간 삭제 refactor: 테스트용(Emulator) 로직 및 쉘스크립트 삭제
