+++ emulator/src/main/java/com/pickcar/domain/CycleInfo.java
@@ -26,7 +26,6 @@ public class CycleInfo {
 
     private Long carId;
 
-
     private LocalDateTime occurredAt;
 
     private Integer cycleCnt;
